src/multimodal-embeddings/requirements.txt
transformers
pillow
scipy
torch
requests
numpy
pandas
seaborn
matplotlib
tqdm
torchvision
scikit-learn
memory_profiler
setuptools
librosa
clip
